m60: Add /init script for loading another Linux
This adds a script for switching the root to another partition, to boot a full-fledged
distribution like Ubuntu installed in the eMMC flash, and enables ext4 for using binary
images.

To use, interrupt the u-boot countdown and edit your bootargs to include a valid root=
device path, and rdinit=/switch to activate the script.  You must specify root= with a
full device path, LABEL= and UUID= are not currently supported.